+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### 1.0.x
 
-* Added:    User model validation to ensure that the identifier is not used by any other active record.
-* Removed:  Restriction on changing identifier column value after model is created
-* Changed:  LoginProvider uses this modules own AuthenticationSettings object to determine the default identity column.
+* Added:        Method to load users from an identity column value
+* Added:        User model validation to ensure that the identifier is not used by any other active record.
+* Removed:      Restriction on changing identifier column value after model is created
+* Changed:      LoginProvider uses this modules own AuthenticationSettings object to determine the default identity column.
 
 ### 1.0.1